Corrige le calcul de la charge + ebauche de load balancer + meilleur calcul de fib(n)

+++ Client/src/main/java/edu/m2dl/s10/arge/openstack/client/Client.java
@@ -93,8 +93,8 @@ public class Client {
             try {
                 Integer random = randInt(10, 100);
                 Object[] params = new Object[] {random};
-                Double result = null;
-                result = (Double) client.execute("Repartiteur.request", params);
+                Long result = null;
+                result = (Long) client.execute("Repartiteur.request", params);
                 System.out.println("fib("+random+") = "+result);
             } catch (XmlRpcException e) {
                 e.printStackTrace();
